feat(experimentalIdentityAndAuth): reorganize package structure into NPM packages
