+++ .changeset/big-panthers-collect.md
@@ -1,0 +1,5 @@
+---
+"@smithy/experimental-identity-and-auth": patch
+---
+
+Reorganize package structure into NPM packages.
