+++ packages/experimental-identity-and-auth/src/index.ts
@@ -1,14 +1,8 @@
-export * from "./HttpAuthScheme";
-export * from "./HttpAuthSchemeProvider";
-export * from "./HttpSigner";
-export * from "./IdentityProviderConfig";
-export * from "./SigV4Signer";
-export * from "./apiKeyIdentity";
-export * from "./endpointRuleSet";
-export * from "./httpApiKeyAuth";
-export * from "./httpBearerAuth";
-export * from "./memoizeIdentityProvider";
 export * from "./middleware-http-auth-scheme";
 export * from "./middleware-http-signing";
-export * from "./noAuth";
-export * from "./tokenIdentity";
+export * from "./signature-v4";
+export * from "./types";
+export * from "./util-endpoint-rule-set-auth";
+export * from "./util-http-api-key-auth";
+export * from "./util-http-bearer-auth";
+export * from "./util-identity-and-auth";
