+++ packages/experimental-identity-and-auth/src/middleware-http-signing/httpSigningMiddleware.ts
@@ -9,7 +9,7 @@ import {
 } from "@smithy/types";
 import { getSmithyContext } from "@smithy/util-middleware";
 
-import { SelectedHttpAuthScheme } from "../HttpAuthScheme";
+import { SelectedHttpAuthScheme } from "../types/auth/HttpAuthScheme";
 
 /**
  * @internal
