+++ packages/experimental-identity-and-auth/src/signature-v4/index.ts
@@ -1,0 +1,1 @@
+export * from "./SigV4Signer";
