+++ packages/experimental-identity-and-auth/src/types/auth/HttpAuthSchemeProvider.ts
