+++ packages/experimental-identity-and-auth/src/types/auth/HttpSigner.ts
