+++ packages/experimental-identity-and-auth/src/types/auth/index.ts
@@ -1,0 +1,3 @@
+export * from "./HttpAuthScheme";
+export * from "./HttpAuthSchemeProvider";
+export * from "./HttpSigner";
