+++ packages/experimental-identity-and-auth/src/util-endpoint-rule-set-auth/index.ts
@@ -1,0 +1,1 @@
+export * from "./endpointRuleSet";
