+++ packages/experimental-identity-and-auth/src/util-http-api-key-auth/index.ts
@@ -1,0 +1,1 @@
+export * from "./httpApiKeyAuth";
