+++ packages/experimental-identity-and-auth/src/util-http-bearer-auth/index.ts
@@ -1,0 +1,1 @@
+export * from "./httpBearerAuth";
