+++ packages/experimental-identity-and-auth/src/util-identity-and-auth/memoizeIdentityProvider.ts
